code/back/requirements.txt
django
djangorestframework
pymysql
cryptography